+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Return wrapped error type from json parsing rules
 - Remove faulty check for valid identifier for rules names
+- Simplified index expressions so you can only index by string or usize, not by expressions anymore
 
 ## [0.6.6] - 2023-07-03